Revisi DB

+++ database/migrations/2024_05_20_113648_create_pembayaran_iuran_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('pembayaran_iuran', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('id_warga');
-            $table->unsignedInteger('id_iuran');
+            $table->unsignedBigInteger('id_warga');
+            $table->unsignedBigInteger('id_iuran');
             $table->date('tanggal_pembayaran');
             $table->integer('jumlah_pembayaran');
             // Define foreign key constraints
